The pos variable needs to be size_t rather then an unsigned int to ensure that its type matches string::npos to avoid an infinite loop here.  Fixes #65.

+++ src/components/cpp/ServiceDirectory/ServiceDirectory.cpp
@@ -138,9 +138,8 @@ static bool validateDomainName(string domain)
 static int parseDomainCSV(string &csv)
 {
 	int prevPos=0;
-	unsigned int pos = csv.find(",",0);
+	size_t pos = csv.find(",",0);
 	int commaCount = 0;
-
 	if(csv.length()==0)
 	{
 		return 0;
@@ -1012,7 +1011,6 @@ void ServiceDirectory::publishDomainUpdateMessage(string updateDomain, string ur
 	GravityDataProduct gdp("ServiceDirectory_DomainUpdate");
 	gdp.setData(updatePB);
 	gn.publish(gdp);
-
 }
 
 } /* namespace gravity */
